Extract out list nested prefix function and strip trailing delimiter
